refactor(tool): split ast tool into modules, add typed schema, address review
Split the ast tool into mode-focused files (outline, symbols, query, naming).
Register a typed ast tool schema so the model receives mode/query/line/depth/
allowIgnored params. Dedup allowIgnored JSON key into a shared constant.
Address review feedback: simplify firstLine, strengthen node ancestry assertion,
and consolidate mode validation into a table-driven test.
